Move DocumentType to persistence package
